+++ src/Haskoin/Transaction.hs
@@ -12,6 +12,7 @@ module Haskoin.Transaction (
     module Common,
     module Builder,
     module Segwit,
+    module Taproot,
     module Partial,
 ) where
 
@@ -19,3 +20,4 @@ import Haskoin.Transaction.Builder as Builder
 import Haskoin.Transaction.Common as Common
 import Haskoin.Transaction.Partial as Partial
 import Haskoin.Transaction.Segwit as Segwit
+import Haskoin.Transaction.Taproot as Taproot